Split ExternalFileScanNode into FileQueryScanNode and FileLoadScanNode. Remove some useless code in FileLoadScanNode.
