+++ fe/fe-core/src/main/java/org/apache/doris/planner/RuntimeFilterGenerator.java
@@ -28,7 +28,7 @@ import org.apache.doris.analysis.TupleId;
 import org.apache.doris.catalog.Type;
 import org.apache.doris.common.IdGenerator;
 import org.apache.doris.common.util.BitUtil;
-import org.apache.doris.planner.external.ExternalFileScanNode;
+import org.apache.doris.planner.external.FileQueryScanNode;
 import org.apache.doris.qe.ConnectContext;
 import org.apache.doris.qe.SessionVariable;
 import org.apache.doris.thrift.TRuntimeFilterMode;
@@ -379,7 +379,7 @@ public final class RuntimeFilterGenerator {
      * 2. Only olap scan nodes are supported:
      */
     private void assignRuntimeFilters(ScanNode scanNode) {
-        if (!(scanNode instanceof OlapScanNode) && !(scanNode instanceof ExternalFileScanNode)) {
+        if (!(scanNode instanceof OlapScanNode) && !(scanNode instanceof FileQueryScanNode)) {
             return;
         }
         TupleId tid = scanNode.getTupleIds().get(0);
