+++ fe/fe-core/src/main/java/org/apache/doris/planner/external/iceberg/IcebergSource.java
@@ -23,7 +23,7 @@ import org.apache.doris.common.DdlException;
 import org.apache.doris.common.MetaNotFoundException;
 import org.apache.doris.common.UserException;
 import org.apache.doris.datasource.ExternalCatalog;
-import org.apache.doris.planner.external.ExternalFileScanNode;
+import org.apache.doris.planner.external.FileQueryScanNode;
 import org.apache.doris.thrift.TFileAttributes;
 
 public interface IcebergSource {
@@ -32,7 +32,7 @@ public interface IcebergSource {
 
     org.apache.iceberg.Table getIcebergTable() throws MetaNotFoundException;
 
-    ExternalFileScanNode.ParamCreateContext createContext() throws UserException;
+    FileQueryScanNode.ParamCreateContext createContext() throws UserException;
 
     TableIf getTargetTable();
